chore: fix all clippy warnings (v0.2.38)
- Remove redundant `as u32` casts on MALLOC_CAP_INTERNAL
- Replace manual dangling pointers with std::ptr::dangling_mut()
- Rename pressure_history::from_bytes → load_from_bytes (from_* convention)
- Use matches!() macro for Ok(Some(_)) patterns
- Use is_multiple_of() for % == 0 checks
- Use (1..=12).contains() for range check
- Collapse nested if into single condition in render loop
- Suppress too_many_arguments on draw_line_graph (intentional)

+++ src/console.rs
@@ -111,7 +111,7 @@ fn process_line(
             info!("uptime: {}h {}m", hours, mins);
             let heap_kb = unsafe { esp_idf_sys::esp_get_free_heap_size() } / 1024;
             info!("free heap: {} KB", heap_kb);
-            let sram_kb = unsafe { esp_idf_sys::heap_caps_get_largest_free_block(esp_idf_sys::MALLOC_CAP_INTERNAL as u32) } / 1024;
+            let sram_kb = unsafe { esp_idf_sys::heap_caps_get_largest_free_block(esp_idf_sys::MALLOC_CAP_INTERNAL) } / 1024;
             info!("SRAM block: {} KB", sram_kb);
             info!("hint: type 'help' for all commands");
         }
